Initialization of Embedding::geom and ::ghost in NCMesh and Mesh.
