+++ mesh/ncmesh.cpp
@@ -1865,8 +1865,12 @@ void NCMesh::InitDerefTransforms()
    transforms.embeddings.SetSize(nfine);
    for (int i = 0; i < nfine; i++)
    {
-      transforms.embeddings[i].parent = -1;
-      transforms.embeddings[i].matrix = 0;
+      Embedding &emb = transforms.embeddings[i];
+      emb.parent = -1;
+      emb.matrix = 0;
+      Element &el =elements[leaf_elements[i]];
+      emb.geom = el.Geom();
+      emb.ghost = IsGhost(el);
    }
 }
 
@@ -1879,7 +1883,7 @@ void NCMesh::SetDerefMatrixCodes(int parent, Array<int> &fine_coarse)
       Element &ch = elements[prn.child[i]];
       if (ch.index >= 0)
       {
-         int code = (prn.ref_type << 8) | (i << 4) | prn.geom;
+         int code = (prn.ref_type << 4) | i;
          transforms.embeddings[ch.index].matrix = code;
          fine_coarse[ch.index] = parent;
       }
@@ -4291,6 +4295,8 @@ void NCMesh::TraverseRefinements(int elem, int coarse_index,
       Embedding &emb = transforms.embeddings[el.index];
       emb.parent = coarse_index;
       emb.matrix = matrix - 1;
+      emb.geom = el.Geom();
+      emb.ghost = IsGhost(el);
    }
    else
    {
@@ -4378,17 +4384,14 @@ const CoarseFineTransformations& NCMesh::GetDerefinementTransforms()
       // assign numbers to the different matrices used
       for (int i = 0; i < transforms.embeddings.Size(); i++)
       {
-         int code = transforms.embeddings[i].matrix;
+         Embedding &emb = transforms.embeddings[i];
+         int code = emb.matrix; // see SetDerefMatrixCodes()
          if (code)
          {
-            int geom = code & 0xf; // see SetDerefMatrixCodes()
-            int ref_type_child = code >> 4;
+            int &matrix = mat_no[emb.geom][code];
+            if (!matrix) { matrix = mat_no[emb.geom].size(); }
 
-            int &matrix = mat_no[geom][ref_type_child];
-            if (!matrix) { matrix = mat_no[geom].size(); }
-
-            transforms.embeddings[i].matrix = matrix - 1;
-            transforms.embeddings[i].geom = geom; // TODO
+            emb.matrix = matrix - 1;
          }
       }
 
@@ -4422,36 +4425,6 @@ const CoarseFineTransformations& NCMesh::GetDerefinementTransforms()
    }
    return transforms;
 }
-
-namespace internal
-{
-
-// Used in CoarseFineTransformations::GetCoarseToFineMap() below.
-struct RefType
-{
-   Geometry::Type geom;
-   int num_children;
-   const Pair<int,int> *children;
-
-   RefType(Geometry::Type g, int n, const Pair<int,int> *c)
-      : geom(g), num_children(n), children(c) { }
-
-   bool operator<(const RefType &other) const
-   {
-      if (geom < other.geom) { return true; }
-      if (geom > other.geom) { return false; }
-      if (num_children < other.num_children) { return true; }
-      if (num_children > other.num_children) { return false; }
-      for (int i = 0; i < num_children; i++)
-      {
-         if (children[i].one < other.children[i].one) { return true; }
-         if (children[i].one > other.children[i].one) { return false; }
-      }
-      return false; // everything is equal
-   }
-};
-
-} // namespace internal
 
 void CoarseFineTransformations::MakeCoarseToFineTable(Table &coarse_to_fine,
                                                       bool want_ghosts) const
@@ -4515,63 +4488,6 @@ void CoarseFineTransformations::MakeCoarseToFineTable(Table &coarse_to_fine,
    }
 }
 
-
-/*{
-   coarse_to_ref_type.SetSize(coarse_ne);
-
-   using internal::RefType;
-   using std::map;
-   using std::pair;
-
-   map<RefType,int> ref_type_map;
-   for (int i = 0; i < coarse_ne; i++)
-   {
-      const int num_children = cf_i[i+1]-cf_i[i];
-      MFEM_ASSERT(num_children > 0, "");
-      const int fine_el = cf_j[cf_i[i]].two;
-      // Assuming the coarse and the fine elements have the same geometry:
-      const Geometry::Type geom = new_mesh.GetElementBaseGeometry(fine_el);
-      const RefType ref_type(geom, num_children, &cf_j[cf_i[i]]);
-      pair<map<RefType,int>::iterator,bool> res =
-         ref_type_map.insert(
-            pair<const RefType,int>(ref_type, (int)ref_type_map.size()));
-      coarse_to_ref_type[i] = res.first->second;
-   }
-
-   ref_type_to_matrix.MakeI((int)ref_type_map.size());
-   ref_type_to_geom.SetSize((int)ref_type_map.size());
-   for (map<RefType,int>::iterator it = ref_type_map.begin();
-        it != ref_type_map.end(); ++it)
-   {
-      ref_type_to_matrix.AddColumnsInRow(it->second, it->first.num_children);
-      ref_type_to_geom[it->second] = it->first.geom;
-   }
-
-   ref_type_to_matrix.MakeJ();
-   for (map<RefType,int>::iterator it = ref_type_map.begin();
-        it != ref_type_map.end(); ++it)
-   {
-      const RefType &rt = it->first;
-      for (int j = 0; j < rt.num_children; j++)
-      {
-         ref_type_to_matrix.AddConnection(it->second, rt.children[j].one);
-      }
-   }
-   ref_type_to_matrix.ShiftUpI();
-}*/
-
-/*void CoarseFineTransformations::GetCoarseToFineMap(const Mesh &new_mesh,
-                                                   Table &coarse_to_fine) const
-{
-   Array<int> coarse_to_ref_type;
-   Table ref_type_to_matrix;
-   Array<mfem::Geometry::Type> ref_type_to_geom;
-   bool get_coarse_to_fine_only = true;
-   GetCoarseToFineMap(new_mesh, coarse_to_fine, coarse_to_ref_type,
-                      ref_type_to_matrix, ref_type_to_geom,
-                      get_coarse_to_fine_only);
-}*/
-
 void NCMesh::ClearTransforms()
 {
    coarse_elements.DeleteAll();
@@ -4599,7 +4515,7 @@ bool CoarseFineTransformations::IsInitialized() const
 
 void Swap(CoarseFineTransformations &a, CoarseFineTransformations &b)
 {
-   for (int g=0; g<Geometry::NumGeom; ++g)
+   for (int g = 0; g < Geometry::NumGeom; ++g)
    {
       a.point_matrices[g].Swap(b.point_matrices[g]);
    }
